rural-healthcare-bot/requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
faiss-cpu
pypdf
sqlalchemy
groq
requests
python-dotenv
pydantic
